Alloc Instrumentation in Shmem
On postmaster start, additional space in Shmem is allocated for Instrumentation slots and a header.
The number of slots is controlled by a cluster level GUC, default is 30000 (means 5MB memory).
The default number is estimated by 250 concurent queries * 120 nodes per query.
If the slots are exhausted, instruments are allocated in local memory as fallback.
From monitoring prespective, it is acceptable that some running node's instrument are not reachable on some rare situation.

These slots are organized as a free list
  - Header points to the first free slot.
  - Each free slot points to next free slot.
  - The last free slot's next pointer is NULL.
When GUC gp_enable_query_metrics is on, InstrShemmPick called from ExecInitNode will pick an empty slot from the free list to store Instrumentation.
  - The free slot pointed by the header is picked.
  - The picked slot's next pointer is assigned to the header.
  - A spin lock on the header to prevent concurrently writing.
  - When the GUC is off, Instrumentation will still be allocated in local memory.
On ExecEndNode or query abort/error (mppExecutorCleanup), the occupied slot is cleaned up and returned to the free list
  - The returning slot is wiped with all zero.
  - The returning slot's next pointer is set to the current value of header.
  - The header is pointed to the returning slot.
  - A spin lock is also used to protect header.
On FATAL happends, backend will exit with proc_exit(1) without mppExecutorCleanup,
  - To prevent leak, recycle all solts used by this process in proc_exit_prepare -> shmem_exit -> CleanupCancelBackend.
  - Other backends serving same query will recieve cancel signal and tear down with mppExecutorCleanup.